+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_text_analytics.py
@@ -34,9 +34,12 @@ class TextAnalyticsTest(TestAnalyticsTestCase):
 
         text_document_statistics = _models.TextDocumentStatistics(grapheme_count=14, transaction_count=18)
 
+        warnings = [_models.TextAnalyticsWarning(code="LongWordsInDocument", message="The document contains very long words (longer than 64 characters). These words will be truncated and may result in unreliable model predictions.")]
+
         recognize_entities_result = _models.RecognizeEntitiesResult(
             id="1",
             entities=[categorized_entity],
+            warnings=warnings,
             statistics=text_document_statistics,
             is_error=False
         )
@@ -44,6 +47,7 @@ class TextAnalyticsTest(TestAnalyticsTestCase):
         detect_language_result = _models.DetectLanguageResult(
             id="1",
             primary_language=detected_language,
+            warnings=warnings,
             statistics=text_document_statistics,
             is_error=False
         )
@@ -56,7 +60,7 @@ class TextAnalyticsTest(TestAnalyticsTestCase):
 
         extract_key_phrases_result = \
             _models.ExtractKeyPhrasesResult(
-                id="1", key_phrases=["dog", "cat", "bird"], statistics=text_document_statistics, is_error=False
+                id="1", key_phrases=["dog", "cat", "bird"], warnings=warnings, statistics=text_document_statistics, is_error=False
             )
 
         linked_entity_match = _models.LinkedEntityMatch(confidence_score=0.999, text="Bill Gates", grapheme_offset=0,
@@ -72,7 +76,7 @@ class TextAnalyticsTest(TestAnalyticsTestCase):
         )
         recognize_linked_entities_result = \
             _models.RecognizeLinkedEntitiesResult(
-                id="1", entities=[linked_entity], statistics=text_document_statistics, is_error=False
+                id="1", entities=[linked_entity], warnings=warnings, statistics=text_document_statistics, is_error=False
             )
 
         sentiment_confidence_score_per_label = \
@@ -83,13 +87,13 @@ class TextAnalyticsTest(TestAnalyticsTestCase):
             sentiment="neutral",
             confidence_scores=sentiment_confidence_score_per_label,
             grapheme_offset=0,
-            grapheme_length=10,
-            warnings=["sentence was too short to find sentiment"]
+            grapheme_length=10
         )
 
         analyze_sentiment_result = _models.AnalyzeSentimentResult(
             id="1",
             sentiment="positive",
+            warnings=warnings,
             statistics=text_document_statistics,
             confidence_scores=sentiment_confidence_score_per_label,
             sentences=[sentence_sentiment],
@@ -117,15 +121,22 @@ class TextAnalyticsTest(TestAnalyticsTestCase):
                          repr(text_document_statistics))
         self.assertEqual("RecognizeEntitiesResult(id=1, entities=[CategorizedEntity(text=Bill Gates, category=Person, "
                          "subcategory=Age, grapheme_offset=0, grapheme_length=8, confidence_score=0.899)], "
+                         "warnings=[TextAnalyticsWarning(code=LongWordsInDocument, message=The document contains very long words (longer than 64 characters). "
+                         "These words will be truncated and may result in unreliable model predictions.)], "
                          "statistics=TextDocumentStatistics(grapheme_count=14, transaction_count=18), "
                          "is_error=False)", repr(recognize_entities_result))
         self.assertEqual("DetectLanguageResult(id=1, primary_language=DetectedLanguage(name=English, "
-                         "iso6391_name=en, confidence_score=1.0), statistics=TextDocumentStatistics(grapheme_count=14, "
+                         "iso6391_name=en, confidence_score=1.0), "
+                         "warnings=[TextAnalyticsWarning(code=LongWordsInDocument, message=The document contains very long words (longer than 64 characters). "
+                         "These words will be truncated and may result in unreliable model predictions.)], "
+                         "statistics=TextDocumentStatistics(grapheme_count=14, "
                          "transaction_count=18), is_error=False)", repr(detect_language_result))
         self.assertEqual("TextAnalyticsError(code=invalidRequest, message=The request is invalid, target=request)",
                          repr(text_analytics_error))
-        self.assertEqual("ExtractKeyPhrasesResult(id=1, key_phrases=['dog', 'cat', 'bird'], statistics="
-                         "TextDocumentStatistics(grapheme_count=14, transaction_count=18), is_error=False)",
+        self.assertEqual("ExtractKeyPhrasesResult(id=1, key_phrases=['dog', 'cat', 'bird'], "
+                         "warnings=[TextAnalyticsWarning(code=LongWordsInDocument, message=The document contains very long words (longer than 64 characters). "
+                         "These words will be truncated and may result in unreliable model predictions.)], "
+                         "statistics=TextDocumentStatistics(grapheme_count=14, transaction_count=18), is_error=False)",
                          repr(extract_key_phrases_result))
         self.assertEqual("LinkedEntityMatch(confidence_score=0.999, text=Bill Gates, grapheme_offset=0, grapheme_length=8)",
                          repr(linked_entity_match))
@@ -138,18 +149,25 @@ class TextAnalyticsTest(TestAnalyticsTestCase):
                          "grapheme_length=8), LinkedEntityMatch(confidence_score=0.999, text=Bill Gates, grapheme_offset=0, "
                          "grapheme_length=8)], language=English, data_source_entity_id=Bill Gates, "
                          "url=https://en.wikipedia.org/wiki/Bill_Gates, data_source=wikipedia)], "
+                         "warnings=[TextAnalyticsWarning(code=LongWordsInDocument, message=The document contains very long words (longer than 64 characters). "
+                         "These words will be truncated and may result in unreliable model predictions.)], "
                          "statistics=TextDocumentStatistics(grapheme_count=14, "
                          "transaction_count=18), is_error=False)", repr(recognize_linked_entities_result))
         self.assertEqual("SentimentConfidenceScores(positive=0.99, neutral=0.05, negative=0.02)",
                          repr(sentiment_confidence_score_per_label))
         self.assertEqual("SentenceSentiment(text=This is a sentence., sentiment=neutral, confidence_scores=SentimentConfidenceScores("
-                         "positive=0.99, neutral=0.05, negative=0.02), grapheme_offset=0, grapheme_length=10)", repr(sentence_sentiment))
-        self.assertEqual("AnalyzeSentimentResult(id=1, sentiment=positive, statistics=TextDocumentStatistics("
+                         "positive=0.99, neutral=0.05, negative=0.02), grapheme_offset=0, grapheme_length=10)",
+                         repr(sentence_sentiment))
+        self.assertEqual("AnalyzeSentimentResult(id=1, sentiment=positive, "
+                         "warnings=[TextAnalyticsWarning(code=LongWordsInDocument, message=The document contains very long words (longer than 64 characters). "
+                         "These words will be truncated and may result in unreliable model predictions.)], "
+                         "statistics=TextDocumentStatistics("
                          "grapheme_count=14, transaction_count=18), confidence_scores=SentimentConfidenceScores"
                          "(positive=0.99, neutral=0.05, negative=0.02), "
                          "sentences=[SentenceSentiment(text=This is a sentence., sentiment=neutral, confidence_scores="
                          "SentimentConfidenceScores(positive=0.99, neutral=0.05, negative=0.02), "
-                         "grapheme_offset=0, grapheme_length=10)], is_error=False)",
+                         "grapheme_offset=0, grapheme_length=10)], "
+                         "is_error=False)",
                          repr(analyze_sentiment_result))
         self.assertEqual("DocumentError(id=1, error=TextAnalyticsError(code=invalidRequest, "
                          "message=The request is invalid, target=request), is_error=True)", repr(document_error))