+++ sdk/textanalytics/azure-ai-textanalytics/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0b5 (Unreleased)
 
 **New features**
+- We now have a `warnings` property on each document-level response object returned from the endpoints. It is a list of `TextAnalyticsWarning`s.
 - Added `text` property to `SentenceSentiment`
 
 **Breaking changes**